Implemented function _identify_SixJSymbol()
Given a set() containing four 3j-symbols the function will identify the
corresponding 6j-symbol, and raise exceptions if it is considered impossible.
This is intended to be a helper function for the refine_tjs2sjs() driver
routine, and it will not accept a sympy expression as input.

identify_SixJSymbol() is able to identify all permutations of 6j-symbols
when the 3j-symbols were generated by SixJSymbol.get_ito_ThreeJSymbols()

Added Exception ThreeJSymbolsNotCompatibleWithSixJSymbol

Made stubs for
	- refine_tjs2sjs(expr)
	- _iter_tjs_permutations(expr)