Fix a subtle error by enforcing: -/+symbol <=> antiparticle/particle
Sympy requires that all objects are uniquely defined by __class__ and
self.args.  This is elsewhere stated as the assumption:

	obj.func(*obj.args) == obj

Where Basic.func is mapped to obj.__class__.  The old **kw_args treatment in
QuantumState.__new__() to determine particle/antiparticle meant that whenever
a quantum state 'a' was recreated with 'a=a.func(*a.args)', the **kw_args was
lost.  Antiparticles could unexpectedly turn into particles this way.

From now, all states with -symbol are antiparticles, and else they are
particles.  This was until now used to differentiate output, as antiparticles
were denoted with a negative sign.  Now that convention is enforced as
the only way particles can be differentiated from antiparticles.

 - The method .is_hole is now a property that checks the coefficient of
   self.symbol
 - QuantumState.__new__() makes sure that a string symbol '-a' is interpreted
   as -Symbol('a'), i.e. as an antiparticles.
 - QuantumState.__new__() will still accept the keyword 'hole', and will
   invert the sign of the given symbol if hole=True.