First implementation of an autowrap utility
The module provides a friendly function that is able to compile and wrap a
sympy expression, relying on the codegen utility and wrapper backends.  The
binary code is made available as a regular python function.  The following
doctest works already:

    >>> from sympy.abc import x, y, z
    >>> from sympy.utilities.autowrap import autowrap
    >>> expr = ((x - y + z)**(13)).expand()
    >>> binary_func = autowrap(expr)
    >>> binary_func(1, 4, 2)
    -1.0

The autowrap module provides a ``one-button'' interface that hides all
technical details about wrapping binary code for use with python. Currently,
the only available backend is f2py, but others can easily be added.