autowrap: new method binary_function()
While autowrap() returns the binary code as a callable python function,
the new method binary_function() returns a sympy.core.Function object
with a binary callable stored as the _imp_ attribute.  The new routine
is a very thin wrapper that relies on autowrap() and
implemented_function() from the lambdify module.