ccode, fcode: Add support for inline function printing
The initialization of array elements from a sympy expression requires
loops like

do i=1, n
   y[i] = some_expression_with_outer_index_i
end do

Because every Indexed in some_expression_with_outer_index_i are subject
to contraction, some initializations used to be impossible without a function
call in the innermost loop.  For instance, to define each element in y
in terms of an element in x according to the formula

	y[i] = x[i]*(x[i] + 1)

The right hand side will be interpreted as a broadcast of 1 and a
contraction over i.  The old way to work around this was to use a call
to a (scalar) function f(x) := x*(x+1).  The code printers would then
see the sympy expression as

	y[i] = f(x[i])

so there would be no unintended contractions. This works, but the code
will often get a severe performance penalty.

This patch enables the code printers to inline the expression, by
checking the _imp_ attribute of unknown functions.  If it is defined
and refers to a sympy.core.Lambda instance, the expression stored as the
Lambda is printed instead of the function call.

Added tests for ccode and fcode