secondquant: Do not raise exception for non-dummy index
The normal order class NO, used to refuse the splitting of an index into
below and above fermi if the index was not a summation index. (Of type
Dummy)  The splitting should also be allowed for a non-dummy index, as a
Kronecker delta function will ensure that only one of the terms can be
nonzero for any given value of the index.

The practical benefit is that one can also use wicks() on a NO-object
containing operators that are not subject to summation.