Fix error in powsimp (wrong result)
Before this patch I got:

	>>> from sympy import Pow, powsimp, Symbol
	>>> x = Symbol('x')
	>>> powsimp(Pow(-1,2*x))
	1

This is wrong unless 'x' is known to be an integer, so that the exponent
is even.

The error happens in the simplification 2**(2*x) -> 4**x.  This patch
makes sure that the simplifaction is only applied if the base is
non-negative, and a corresponding test is added.